Implemented m_bar and m200m differences
Changed the way we split the haloes in the model. Added total baryonic
mass for density.Profile class and the actual halo mass for the
component.Component class.

Updated all the halo/model modules and classes to be consistent and
updated the obs_fit calls to these modules.

Also removed all m_fn or hmf references in the code and implemented some
stylistic unity in the method calls for the component.Component and
power.Power classes.