Simplified the model. Removed component
Made the model more physical and representing the layout of the paper. A
power object simply needs a density profile, which is the sum of
different components. The halo mass function change is automatically
calculated from the mass deficit at the provided r200m_inf.

There are some changes with the old version that cause the match with
observations to be less good. This is because of the removal of the
f_comp**2 contribution in the 1h term. This still needs to be
investigated.