Added test in c200m interpolation to prevent crashing
Not closing the files results in the OSError 24, too many open files
when converting halo masses, which requires many calls to the
interpolation. Opening the files with a context handler, on the other
hand, seemed to fail since the contents are memmapped and closed
before they are actually accessed. Hence, we tried to copy the array
contents when opening the file.